refactor: move DiagnosticsAdapter to new file
